+++ packages/dso-toolkit/src/components/list-button/list-button.args.ts
@@ -15,7 +15,6 @@ export interface ListButtonArgs {
   count?: number;
   min?: number;
   max?: number;
-  manual: boolean;
   dsoCountChange: HandlerFunction;
   dsoSelectedChange: HandlerFunction;
 }
@@ -64,11 +63,6 @@ export const listButtonArgTypes: ArgTypes<ListButtonArgs> = {
   max: {
     control: {
       type: "number",
-    },
-  },
-  manual: {
-    control: {
-      type: "boolean",
     },
   },
   dsoCountChange: {